move getTotalSavings to subtract upcoming from available
