+++ src/extension/features/budget/subtract-upcoming-from-available/index.js
@@ -1,8 +1,8 @@
 import { Feature } from 'toolkit/extension/features/feature';
 import { addToolkitEmberHook } from 'toolkit/extension/utils/toolkit';
-import { shouldInvoke } from 'toolkit/extension/features/budget/subtract-upcoming-from-available/util';
 import { handleBudgetBreakdown } from 'toolkit/extension/features/budget/subtract-upcoming-from-available/budget-breakdown';
 import { handleBudgetTableRow } from 'toolkit/extension/features/budget/subtract-upcoming-from-available/budget-table-row';
+import { shouldInvoke } from 'toolkit/extension/features/budget/subtract-upcoming-from-available/util';
 
 export class SubtractUpcomingFromAvailable extends Feature {
   invoke() {
